Fixing dataloader sorting issue (#7)
Prior to this commit, the files lists were sorted alphabetically (according to the filename) and the reference voices were sorted numerically (according to the index of the file id). This was leading to mismatching between pairs of files and reference voices.

This commit fixes the issue by sorting the files according to its fileid not by the whole filename. This only happens when the file names matches the pattern of `fileid_xxx`. It also fixes an issue where the psn list was always empty.