DNSMON support for PyTorch (#9)
This commit converts the DNSMOS from using onnx runtime into using a pytorch model. This makes the whole flow runs on pytorch and makes it possible to run the metric on gpu.

It also fixes an issue with the base model where the metrics and losses always runs on the cpu because of not using ModuleDict.