Added testing support (#17)
This commit adds the support for testing. It adds test.py file that can be used to test any model. It also adds the required configuration files for testing.
It also adds the support for test split in pdns_dataset.py and adds the appropriate configuration for DNS 2023 blind testset.